+++ CHANGES.rst
@@ -6,6 +6,8 @@ Changelog
 
 * Adjust rendering of BitBar plugin errors.
 * Fix fetching of log level names in Python 3.4+.
+* Print CLI output in unittests.
+* Print more debug info in unittests when CLI produce tracebacks.
 
 
 `2.0.0 (2016-12-04) <https://github.com/kdeldycke/meta-package-manager/compare/v1.12.0...v2.0.0>`_